[ApplicationPage] 시술 내역 드롭다운 컴포넌트 분리 (#429)
* refactor: 드롭다운 컴포넌트 분리

* refactor: 컴포넌트 케이스 분리

* refactor: 불필요한 코드 정리
